@@ -27,7 +27,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
         {
             var viewModel = new PageViewModel();
 
-            var authData = await _swiftService.Authenticate();
+            var authData = await _swiftService.AuthenticateAsync();
 
             if (authData != null)
             {
@@ -56,7 +56,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
                     {
                         await fileStream.CopyToAsync(memoryStream);
 
-                        var resp = await _swiftService.PutObjectChunk(containerTempId, fileName, memoryStream.ToArray(), segment);
+                        var resp = await _swiftService.PutObjectChunkAsync(containerTempId, fileName, memoryStream.ToArray(), segment);
 
                         return new JsonResult(new
                         {
@@ -79,10 +79,10 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
         public async Task<IActionResult> UploadDone(int segmentsCount, string fileName, string contentType)
         {
             // use manifest to merge chunks
-            await _swiftService.PutManifest(containerTempId, fileName);
+            await _swiftService.PutManifestAsync(containerTempId, fileName);
 
             // copy chunks to new file and set some meta data info about the file (filename, contentype)
-            await _swiftService.CopyObject(containerTempId, fileName, containerDemoId, fileName, new Dictionary<string, string>
+            await _swiftService.CopyObjectAsync(containerTempId, fileName, containerDemoId, fileName, new Dictionary<string, string>
             {
                 { $"X-Object-Meta-{metaFileName}", fileName },
                 { $"X-Object-Meta-{metaContentType}", contentType }
@@ -99,7 +99,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
 
         public async Task<IActionResult> PlayVideo(string containerId, string objectId)
         {
-            var headObject = await _swiftService.HeadObject(containerId, objectId);
+            var headObject = await _swiftService.HeadObjectAsync(containerId, objectId);
 
             if (headObject.IsSuccess)
             {
@@ -108,7 +108,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
 
                 var stream = new BufferedHTTPStream((start, end) =>
                 {
-                    using (var response = _swiftService.GetObjectRange(containerId, objectId, start, end).Result)
+                    using (var response = _swiftService.GetObjectRangeAsync(containerId, objectId, start, end).Result)
                     {
                         var ms = new MemoryStream();
 
@@ -129,7 +129,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
 
         public async Task<IActionResult> DownloadFile(string containerId, string objectId)
         {
-            var headObject = await _swiftService.HeadObject(containerId, objectId);
+            var headObject = await _swiftService.HeadObjectAsync(containerId, objectId);
 
             if (headObject.IsSuccess && headObject.ContentLength > 0)
             {
@@ -140,7 +140,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
 
                 var stream = new BufferedHTTPStream((start, end) =>
                 {
-                    using (var response = _swiftService.GetObjectRange(containerId, objectId, start, end).Result)
+                    using (var response = _swiftService.GetObjectRangeAsync(containerId, objectId, start, end).Result)
                     {
                         var ms = new MemoryStream();
 
@@ -169,7 +169,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
         {
             var tree = new TreeViewModel();
 
-            var accountData = await _swiftService.GetAccount();
+            var accountData = await _swiftService.GetAccountAsync();
 
             if (accountData.IsSuccess)
             {
@@ -196,7 +196,7 @@ namespace SwiftClient.AspNetCore.Demo.Controllers
 
         private async Task<TreeViewModel> GetContainerBranch(string containerId)
         {
-            var containerData = await _swiftService.GetContainer(containerId);
+            var containerData = await _swiftService.GetContainerAsync(containerId);
 
             TreeViewModel result = new TreeViewModel
             {
